put scatter plot into its own component
